Start making gitdm.py importable
Create two new functions PrintSummary() and PrintReports()

Invoke them conditionally from the usual `if __name__ == "__main__":` to
make `import gitdm` possible.

Once gitdm becomes a library, it's now easy to extract specific datasets
using custom Python scripts (instead of screen-scraping gitdm's output)

The regular diff for this commit is large but it's small when ignoring
whitespace/indentation with `git show -w`.

Warning: `import gitdm` is NOT your usual "import" at all yet. It:
- "hijacks" sys.argv and parses them
- reads and parses the git log from stdin

But it's usable and it's a start!

Signed-off-by: Marc Herbert <marc.herbert@intel.com>